requirements.txt
Django==1.8.2
django-excel
pyexcel-xls
pyexcel-xlsx
django-bootstrap3
eventlet
django-celery
django-cron
django-permissivecsrf
gspread
celery
django-kombu
psycopg2
PyOpenSSL
xlsxwriter
grequests
django-chosen
BeautifulSoup4
python-social-auth